Refactor check into a new package
