+++ scripts/src/java/org/oppia/android/scripts/ci/BUILD.bazel
@@ -16,15 +16,3 @@ kt_jvm_library(
         "//scripts/src/java/org/oppia/android/scripts/common:git_client",
     ],
 )
-
-kt_jvm_library(
-    name = "maven_dependencies_list_check_lib",
-    testonly = True,
-    srcs = [
-        "MavenDependenciesListCheck.kt",
-    ],
-    visibility = ["//scripts:oppia_script_binary_visibility"],
-    deps = [
-        "//scripts/src/java/org/oppia/android/scripts/common:maven_dependencies_list_generator",
-    ],
-)
