+++ scripts/src/java/org/oppia/android/scripts/license/BUILD.bazel
@@ -1,0 +1,15 @@
+"""
+Libraries corresponding to maven dependencies and their licenses checks.
+"""
+
+load("@io_bazel_rules_kotlin//kotlin:kotlin.bzl", "kt_jvm_library")
+
+kt_jvm_library(
+    name = "maven_dependencies_list_check_lib",
+    testonly = True,
+    srcs = ["MavenDependenciesListCheck.kt"],
+    visibility = ["//scripts:oppia_script_binary_visibility"],
+    deps = [
+        "//scripts/src/java/org/oppia/android/scripts/common:maven_dependencies_list_generator",
+    ],
+)
