+++ scripts/src/javatests/org/oppia/android/scripts/license/BUILD.bazel
@@ -1,18 +1,19 @@
 """
-Tests corresponding to developer scripts that help with continuous integration workflows.
+Tests corresponding to the maven dependencies and their license checks.
 """
 
 load("@io_bazel_rules_kotlin//kotlin:kotlin.bzl", "kt_jvm_test")
 
 kt_jvm_test(
-    name = "ComputeAffectedTestsTest",
-    srcs = ["ComputeAffectedTestsTest.kt"],
+    name = "MavenDependenciesListCheckTest",
+    size = "large",
+    srcs = ["MavenDependenciesListCheckTest.kt"],
     deps = [
-        "//scripts/src/java/org/oppia/android/scripts/ci:compute_affected_tests_lib",
+        "//scripts/src/java/org/oppia/android/scripts/license:maven_dependencies_list_check_lib",
         "//scripts/src/java/org/oppia/android/scripts/testing:test_bazel_workspace",
-        "//scripts/src/java/org/oppia/android/scripts/testing:test_git_repository",
         "//testing:assertion_helpers",
         "//third_party:com_google_truth_truth",
         "//third_party:org_jetbrains_kotlin_kotlin-test-junit",
+        "//third_party:org_mockito_kotlin_mockito-kotlin",
     ],
 )
